fixed cursor bug

+++ cpu/ports.c
@@ -6,7 +6,7 @@
     By in, what this really means data from device is coming IN to CPU. This is the euqivalent to reading from device.
     __asm__ function is in AT&T format, so reads left as source and right as destination. The info from port will be in 
     dx, which is then stored in al, which has a C variable assigned to it.*/
-unsigned char port_byte_in(unsigned short port)
+unsigned char port_byte_in(unsigned int port)
 {
 	unsigned char result;
     /*intruction : output : input : clobbered*/
@@ -18,7 +18,7 @@ unsigned char port_byte_in(unsigned short port)
     Out is essentially a write operand. the info we store in al is moved into dx, the sent OUT to device. This is sometimes refered to 
     as a request. You are requesteing data from device by sending data, then reciveing it with IN function
 */
-void port_byte_out(unsigned short port, unsigned char data)
+void port_byte_out(unsigned int port, unsigned char data)
 {
 	__asm__("out %%al, %%dx" : : "a" (data), "d" (port));
 }
